+++ CHANGELOG.md
@@ -14,6 +14,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
   - As an example, the custom stream selector config can specify (in priority order):
     - english audio (and disable subtitles)
     - any other audio (and english subtitles, if they exist)
+  - Criteria can include
+    - Stream language
+    - Stream title (allowed title and/or blocked title)
 
 ### Fixed
 - Fix QSV acceleration in docker with older Intel devices